Decompose integration_tests
