+++ sqlite/functional_tests/integration_tests/src/handlers/key_value.cpp
@@ -1,0 +1,149 @@
+#include "key_value.hpp"
+
+#include <userver/clients/http/component.hpp>
+#include <userver/components/component.hpp>
+#include <userver/components/minimal_server_component_list.hpp>
+#include <userver/formats/serialize/common_containers.hpp>
+#include <userver/server/handlers/http_handler_base.hpp>
+#include <userver/server/handlers/http_handler_json_base.hpp>
+#include <userver/testsuite/testsuite_support.hpp>
+#include <userver/utest/using_namespace_userver.hpp>
+#include <userver/utils/daemon_run.hpp>
+#include "userver/server/http/http_method.hpp"
+
+#include <userver/storages/sqlite/component.hpp>
+#include <userver/storages/sqlite/connection.hpp>
+#include <userver/storages/sqlite/options.hpp>
+#include <userver/storages/sqlite/query.hpp>
+#include <userver/storages/sqlite/result_set.hpp>
+#include <userver/storages/sqlite/transaction.hpp>
+
+#include <db/sql.hpp>
+
+namespace functional_tests {
+
+namespace {
+
+class KeyValue final : public server::handlers::HttpHandlerBase {
+ public:
+  static constexpr std::string_view kName = "handler-key-value";
+
+  KeyValue(const components::ComponentConfig& config,
+           const components::ComponentContext& context)
+      : HttpHandlerBase(config, context),
+        sqlite_connection_(
+            context.FindComponent<components::SQLite>("key-value-database")
+                .GetConnection()) {
+    sqlite_connection_->Execute(db::sql::kCreateTable.data());
+  }
+
+  std::string HandleRequestThrow(const server::http::HttpRequest& request,
+                                 server::request::RequestContext&) const final {
+    const auto& key = request.GetArg("key");
+    if (key.empty()) {
+      throw server::handlers::ClientError(
+          server::handlers::ExternalBody{"No 'key' query argument"});
+    }
+
+    switch (request.GetMethod()) {
+      case server::http::HttpMethod::kGet:
+        return GetValue(key, request);
+      case server::http::HttpMethod::kPost:
+        return PostValue(key, request);
+      case server::http::HttpMethod::kDelete:
+        return DeleteValue(key);
+      case server::http::HttpMethod::kPut:
+        return UpdateValue(key, request);
+      default:
+        throw server::handlers::ClientError(server::handlers::ExternalBody{
+            fmt::format("Unsupported method {}", request.GetMethod())});
+    }
+  }
+
+ private:
+  std::string GetValue(std::string_view key,
+                       const server::http::HttpRequest& request) const {
+    storages::sqlite::ResultSet res =
+        sqlite_connection_->Execute(db::sql::kSelectValueByKey.data(), key);
+    if (res.IsEmpty()) {
+      request.SetResponseStatus(server::http::HttpStatus::kNotFound);
+      return {};
+    }
+
+    return res.AsSingleRow<std::string>();
+  }
+
+  std::string PostValue(std::string_view key,
+                        const server::http::HttpRequest& request) const {
+    const auto& value = request.GetArg("value");
+
+    storages::sqlite::Transaction transaction =
+        sqlite_connection_->Begin("sample_transaction_insert_key_value", {});
+
+    auto res = transaction.Execute(db::sql::kInsertKeyValue.data(), key, value);
+    if (res.RowsAffected()) {
+      transaction.Commit();
+      request.SetResponseStatus(server::http::HttpStatus::kCreated);
+      return std::string{value};
+    }
+
+    res = transaction.Execute(db::sql::kSelectValueByKey.data(), key);
+    transaction.Rollback();
+
+    auto result = res.AsSingleRow<std::string>();
+    if (result != value) {
+      request.SetResponseStatus(server::http::HttpStatus::kConflict);
+    }
+
+    return res.AsSingleRow<std::string>();
+  }
+
+  std::string UpdateValue(std::string_view key,
+                          const server::http::HttpRequest& request) const {
+    const auto& value = request.GetArg("value");
+
+    using storages::sqlite::TransactionOptions;
+
+    storages::sqlite::Transaction transaction = sqlite_connection_->Begin(
+        "sample_transaction_update_key_value",
+        TransactionOptions{TransactionOptions::Mode::kImmediate});
+
+    auto res = sqlite_connection_->Execute(db::sql::kUpdateKeyValue.data(),
+                                           value, key);
+
+    if (res.RowsAffected()) {
+      transaction.Commit();
+      request.SetResponseStatus(server::http::HttpStatus::kCreated);
+      return std::string{value};
+    }
+    res = transaction.Execute(db::sql::kSelectValueByKey.data(), key);
+    transaction.Rollback();
+
+    auto result = res.AsSingleRow<std::string>();
+    if (result != value) {
+      request.SetResponseStatus(server::http::HttpStatus::kConflict);
+    }
+
+    return res.AsSingleRow<std::string>();
+  }
+
+  std::string DeleteValue(std::string_view key) const {
+    const storages::sqlite::Query kDeleteValue{
+        db::sql::kDeleteKeyValue.data(),
+        storages::sqlite::Query::Name{"sample_delete_value"},
+    };
+
+    auto res = sqlite_connection_->Execute(kDeleteValue, key);
+    return std::to_string(res.RowsAffected());
+  }
+
+  storages::sqlite::ConnectionPtr sqlite_connection_;
+};
+
+}  // namespace
+
+void AppendKeyValue(userver::components::ComponentList& component_list) {
+  component_list.Append<KeyValue>();
+}
+
+}  // namespace functional_tests
