+++ sqlite/functional_tests/integration_tests/src/handlers/key_value.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <userver/components/component_list.hpp>
+
+namespace functional_tests {
+
+void AppendKeyValue(userver::components::ComponentList& component_list);
+
+}  // namespace functional_tests
